update setup.sh: changed the execution user from root to regular user

+++ ubuntu/24.04/setup.sh
@@ -1,22 +1,69 @@
 #!/bin/bash
 
-NON_ROOT_USERNAME="ubuntu"
 SSH_AUTHORIZED_KEY_URL="https://raw.githubusercontent.com/nkdgc/server-setup/refs/heads/main/authorized_keys/authorized_keys"
 VIMRC_URL="https://raw.githubusercontent.com/nkdgc/server-setup/refs/heads/main/vim/linux-vimrc"
-BASHRC_URL="https://raw.githubusercontent.com/nkdgc/server-setup/refs/heads/main/bashrc/bashrc.sh"
+BASHRC_URL="https://raw.githubusercontent.com/nkdgc/server-setup/refs/heads/main/bash/dot_bashrc"
+DIRCOLOR_URL="https://raw.githubusercontent.com/nkdgc/server-setup/refs/heads/main/bash/dot_dircolors"
 TMUXCONF_URL="https://raw.githubusercontent.com/nkdgc/server-setup/refs/heads/main/tmux/tmux.conf"
 
-# # Check if the script is running as the root user
-# if [ "$(whoami)" != "root" ]; then
-#     echo "Error: This script must be run as the root user. Please run the script with 'sudo' or as the root user."
-#     exit 1
-# fi
-
-if [ "$#" -lt 1 ] || [ "$#" -gt 2 ]; then
-    echo "Error: This script requires exactly one argument (the hostname), and an optional --silent flag."
-    echo "Usage: $0 <hostname> [--silent]"
+# Check if the script is running as the root user
+if [ "$(whoami)" != "root" ]; then
+    echo "Error: This script must be run as the root user. Please run the script with 'sudo' or as the root user."
     exit 1
 fi
+
+silent=false
+light=false
+
+while (( $# > 0 ))
+do
+  case $1 in
+    # --- silent ---
+    --silent)
+      silent=true
+      ;;
+    # --- light ---
+    --light)
+      light=true
+      ;;
+    # --- hostname ---
+    -h | --hostname | --hostname=*)
+      if [[ "$1" =~ ^--hostname= ]]; then
+        hostname=$(echo $1 | sed -e 's/^--hostname=//')
+      elif [[ -z "$2" ]] || [[ "$2" =~ ^-+ ]]; then
+        echo "'hostname' requires an argument." 1>&2
+        exit 1
+      else
+        hostname="$2"
+        shift
+      fi
+      ;;
+    # --- user ---
+    -h | --user | --user=*)
+      if [[ "$1" =~ ^--user= ]]; then
+        user=$(echo $1 | sed -e 's/^--user=//')
+      elif [[ -z "$2" ]] || [[ "$2" =~ ^-+ ]]; then
+        echo "'user' requires an argument." 1>&2
+        exit 1
+      else
+        user="$2"
+        shift
+      fi
+      ;;
+    # --- other -> ERROR ---
+    *)
+      echo "unknown option: $1" 1>&2
+      exit 1
+      ;;
+  esac
+  shift
+done
+
+echo "----- debug -----"
+echo "silent   : $silent"
+echo "light    : $light"
+echo "hostname : $hostname"
+echo "user     : $user"
 
 function exec_cmd_rc_0(){
   cmd=$1
@@ -32,57 +79,74 @@ function exec_cmd_rc_0(){
 }
 
 # Confirm and change the hostname
-echo "############### change hostname ###############"
-hostname="$1"
-silent_mode=false
-if [ "$#" -eq 2 ] && [ "$2" == "--silent" ]; then
-    silent_mode=true
-fi
-if [ "$silent_mode" = false ]; then
-  read -p "Change the hostname to ${hostname}? (Y/n): " confirm
-  case "$confirm" in
-    [yY]*)
-      echo "Changing hostname to $hostname..."
-      exec_cmd_rc_0 "sudo hostnamectl set-hostname ${hostname}"
-      ;;
-    *)
-      echo "Exiting script."
-      exit 0
-      ;;
-  esac
-else
-    echo "Silent mode enabled. Changing hostname to $hostname..."
-    exec_cmd_rc_0 "sudo hostnamectl set-hostname ${hostname}"
+if [ -n "$hostname" ]; then
+  echo "############### change hostname ###############"
+  if "$silent"; then
+    exec_cmd_rc_0 "hostnamectl set-hostname ${hostname}"
+  else
+    read -p "Change the hostname to ${hostname}? (Y/n): " confirm
+    case "$confirm" in
+      [yY]*)
+        echo "Changing hostname to $hostname..."
+        exec_cmd_rc_0 "hostnamectl set-hostname ${hostname}"
+        ;;
+      *)
+        echo "Exiting script."
+        exit 0
+        ;;
+    esac
+  fi
 fi
 
 echo "############### change timezone ###############"
-exec_cmd_rc_0 "sudo timedatectl set-timezone Asia/Tokyo"
+exec_cmd_rc_0 "timedatectl set-timezone Asia/Tokyo"
 
 echo "############### bashrc ###############"
-exec_cmd_rc_0 "sudo sh -c \"curl ${BASHRC_URL} >> /root/.bashrc\""
-exec_cmd_rc_0 "curl ${BASHRC_URL} >> /home/${NON_ROOT_USERNAME}/.bashrc"
+exec_cmd_rc_0 "curl ${BASHRC_URL} >> ~/.bashrc"
+if [ -n "$user" ]; then
+  exec_cmd_rc_0 "cp ~/.bashrc /home/${user}/.bashrc"
+  exec_cmd_rc_0 "chown ${user}:${user} /home/${user}/.bashrc"
+fi
+
+echo "############### dircolors ###############"
+exec_cmd_rc_0 "curl ${DIRCOLORS_URL} >> ~/.dircolors"
+if [ -n "$user" ]; then
+  exec_cmd_rc_0 "cp ~/.dircolors /home/${user}/.dircolors"
+  exec_cmd_rc_0 "chown ${user}:${user} /home/${user}/.dircolors"
+fi
 
 echo "############### create ssh key ###############"
-exec_cmd_rc_0 "ssh-keygen -t rsa -N '' -f /home/${NON_ROOT_USERNAME}/.ssh/id_rsa"
+exec_cmd_rc_0 "ssh-keygen -t rsa -N '' -f /root/.ssh/id_rsa"
+if [ -n "$user" ]; then
+  exec_cmd_rc_0 "su - ${user} -c \"ssh-keygen -t rsa -N '' -f /home/${user}/.ssh/id_rsa\""
+fi
 
 echo "############### add authorized key ###############"
-exec_cmd_rc_0 "curl ${SSH_AUTHORIZED_KEY_URL} >> /home/${NON_ROOT_USERNAME}/.ssh/authorized_keys"
+exec_cmd_rc_0 "curl ${SSH_AUTHORIZED_KEY_URL} >> ~/.ssh/authorized_keys"
 # delete blank line
-exec_cmd_rc_0 "sed -i '/^$/d' /home/${NON_ROOT_USERNAME}/.ssh/authorized_keys"
-
-# echo "############### change apt repository to ftp.riken.jp"
-# perl -p -i.bak -e 's%(deb(?:-src|)\s+)https?://(?!archive\.canonical\.com|security\.ubuntu\.com)[^\s]+%$1http://ftp.riken.jp/Linux/ubuntu/%' /etc/apt/sources.list
+exec_cmd_rc_0 "sed -i '/^$/d' ~/.ssh/authorized_keys"
+if [ -n "$user" ]; then
+  exec_cmd_rc_0 "cp ~/.ssh/authorized_keys /home/${user}/authorized_keys"
+  exec_cmd_rc_0 "chown ${user}:${user} /home/${user}/authorized_keys"
+fi
 
 echo "############### apt update, upgrade, install vim/git/tmux/... ###############"
-exec_cmd_rc_0 "sudo apt update"
-exec_cmd_rc_0 "sudo dpkg --configure -a"
-exec_cmd_rc_0 "sudo apt upgrade -y"
-exec_cmd_rc_0 "sudo apt install -y vim git traceroute tmux curl net-tools zip unzip postgresql-client apache2-utils"
+exec_cmd_rc_0 "apt update"
+exec_cmd_rc_0 "dpkg --configure -a"
+exec_cmd_rc_0 "apt upgrade -y"
+if "$light"; then
+  exec_cmd_rc_0 "apt install -y vim git traceroute tmux curl net-tools zip unzip postgresql-client apache2-utils"
+else
+  exec_cmd_rc_0 "apt install -y vim git traceroute tmux curl net-tools zip unzip"
+fi
 # memo:
 # - apache2-utils for Apache Bench (ab)
 
 echo "############### setup git ###############"
 exec_cmd_rc_0 "git config --global core.editor vim"
+if [ -n "$user" ]; then
+  exec_cmd_rc_0 "su - ${user} -c \"git config --global core.editor vim\""
+fi
 
 # echo "############### allow ssh root login"
 # SSH_BASE_DIR="/etc/ssh"
@@ -92,14 +156,18 @@ exec_cmd_rc_0 "git config --global core.editor vim"
 # systemctl restart sshd
 
 echo "############### .vimrc ###############"
-exec_cmd_rc_0 "sudo sh -c \"curl ${VIMRC_URL} > /root/.vimrc\""
-exec_cmd_rc_0 "sudo cp /root/.vimrc /home/${NON_ROOT_USERNAME}/.vimrc"
-exec_cmd_rc_0 "sudo chown ${NON_ROOT_USERNAME}:${NON_ROOT_USERNAME} /home/${NON_ROOT_USERNAME}/.vimrc"
+exec_cmd_rc_0 "curl ${VIMRC_URL} > ~/.vimrc"
+if [ -n "$user" ]; then
+  exec_cmd_rc_0 "cp ~/.vimrc /home/${user}/.vimrc"
+  exec_cmd_rc_0 "chown ${user}:${user} /home/${user}/.vimrc"
+fi
 
 echo "############### .tmux.conf ###############"
-exec_cmd_rc_0 "sudo sh -c \"curl ${TMUXCONF_URL} > /root/.tmux.conf\""
-exec_cmd_rc_0 "sudo sh -c \"cp /root/.tmux.conf /home/${NON_ROOT_USERNAME}/.tmux.conf\""
-exec_cmd_rc_0 "sudo sh -c \"chown ${NON_ROOT_USERNAME}:${NON_ROOT_USERNAME} /home/${NON_ROOT_USERNAME}/.tmux.conf\""
+exec_cmd_rc_0 "curl ${TMUXCONF_URL} > ~/.tmux.conf"
+if [ -n "$user" ]; then
+  exec_cmd_rc_0 "cp ~/.tmux.conf /home/${user}/.tmux.conf"
+  exec_cmd_rc_0 "chown ${user}:${user} /home/${user}/.tmux.conf"
+fi
 
 # echo "############### install asciinema"
 # apt-add-repository -y ppa:zanchey/asciinema
@@ -141,21 +209,34 @@ exec_cmd_rc_0 "sudo sh -c \"chown ${NON_ROOT_USERNAME}:${NON_ROOT_USERNAME} /hom
 # 
 # chmod 644 /root/99-config.yaml
 
-echo "############### install docker engine ###############"
-exec_cmd_rc_0 "sudo apt-get update"
-exec_cmd_rc_0 "sudo apt-get install ca-certificates curl gnupg lsb-release"
-exec_cmd_rc_0 "curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo gpg --dearmor -o /usr/share/keyrings/docker-archive-keyring.gpg"
-exec_cmd_rc_0 "echo \"deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/docker-archive-keyring.gpg] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable\" | sudo tee /etc/apt/sources.list.d/docker.list > /dev/null"
-exec_cmd_rc_0 "sudo apt-get update"
-exec_cmd_rc_0 "sudo apt-get install -y docker-ce docker-ce-cli containerd.io docker-compose-plugin"
-exec_cmd_rc_0 "sudo usermod -aG docker ${NON_ROOT_USERNAME}"
-exec_cmd_rc_0 "sudo docker --version"   # Check
+if ! "$light"; then
+  echo "############### install docker engine ###############"
+  exec_cmd_rc_0 "apt-get update"
+  exec_cmd_rc_0 "apt-get install ca-certificates curl gnupg lsb-release"
+  exec_cmd_rc_0 "curl -fsSL https://download.docker.com/linux/ubuntu/gpg | gpg --dearmor -o /usr/share/keyrings/docker-archive-keyring.gpg"
+  exec_cmd_rc_0 "echo \"deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/docker-archive-keyring.gpg] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable\" | tee /etc/apt/sources.list.d/docker.list > /dev/null"
+  exec_cmd_rc_0 "apt-get update"
+  exec_cmd_rc_0 "apt-get install -y docker-ce docker-ce-cli containerd.io docker-compose-plugin"
+  if [ -n "$user" ]; then
+    exec_cmd_rc_0 "usermod -aG docker ${user}"
+  fi
+  exec_cmd_rc_0 "docker --version"   # Check
+fi
 
-echo "############### install aws cli ###############"
-exec_cmd_rc_0 "curl https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip -o awscliv2.zip"
-exec_cmd_rc_0 "unzip awscliv2.zip"
-exec_cmd_rc_0 "sudo ./aws/install"
-exec_cmd_rc_0 "aws --version"           # Check
+if ! "$light"; then
+  echo "############### install aws cli ###############"
+  exec_cmd_rc_0 "curl https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip -o awscliv2.zip"
+  exec_cmd_rc_0 "unzip awscliv2.zip"
+  exec_cmd_rc_0 "./aws/install --bin-dir /usr/local/bin --install-dir /usr/local/aws-cli --update"
+  exec_cmd_rc_0 "aws --version"           # Check
+  if [ -n "$user" ]; then
+    exec_cmd_rc_0 "su - ${user} aws --version"           # Check
+  fi
+fi
+# exec_cmd_rc_0 "curl https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip -o awscliv2.zip"
+# exec_cmd_rc_0 "unzip awscliv2.zip"
+# exec_cmd_rc_0 "sudo ./aws/install"
+# exec_cmd_rc_0 "aws --version"           # Check
 
 # echo "############### install terraform"
 # exec_cmd_rc_0 "wget -O- https://apt.releases.hashicorp.com/gpg | sudo gpg --dearmor -o /usr/share/keyrings/hashicorp-archive-keyring.gpg"
@@ -163,20 +244,41 @@ exec_cmd_rc_0 "aws --version"           # Check
 # exec_cmd_rc_0 "sudo apt update && sudo apt install -y terraform"
 # exec_cmd_rc_0 "terraform -v"            # Check
 
-echo "############### install asdf ###############"
-exec_cmd_rc_0 "git clone https://github.com/asdf-vm/asdf.git ~/.asdf --branch v0.14.1"
-exec_cmd_rc_0 "echo '. "\$HOME/.asdf/asdf.sh"' >> ~/.bashrc"
-exec_cmd_rc_0 '. "$HOME/.asdf/asdf.sh"'
-exec_cmd_rc_0 "echo '. "\$HOME/.asdf/completions/asdf.bash"' >> ~/.bashrc"
+if ! "$light"; then
+  echo "############### install asdf ###############"
+  if [ -z "$user" ]; then
+    exec_cmd_rc_0 "git clone https://github.com/asdf-vm/asdf.git ~/.asdf --branch v0.14.1"
+    exec_cmd_rc_0 "echo '. "\$HOME/.asdf/asdf.sh"' >> ~/.bashrc"
+    exec_cmd_rc_0 '. "$HOME/.asdf/asdf.sh"'
+    exec_cmd_rc_0 "echo '. "\$HOME/.asdf/completions/asdf.bash"' >> ~/.bashrc"
+  else
+    exec_cmd_rc_0 "su - ${user} -c \"git clone https://github.com/asdf-vm/asdf.git ~/.asdf --branch v0.14.1\""
+    exec_cmd_rc_0 "su - ${user} -c \"echo '. "\$HOME/.asdf/asdf.sh"' >> ~/.bashrc\""
+    # exec_cmd_rc_0 'su - ${user} -c . "$HOME/.asdf/asdf.sh"'
+    exec_cmd_rc_0 "su - ${user} -c \"echo '. "\$HOME/.asdf/completions/asdf.bash"' >> ~/.bashrc\""
+  fi
+fi
 
-echo "############### install terraform with asdf ###############"
-exec_cmd_rc_0 "asdf plugin add terraform"
-exec_cmd_rc_0 "asdf list all terraform"
-exec_cmd_rc_0 "asdf install terraform latest"
-exec_cmd_rc_0 "asdf list terraform"
-exec_cmd_rc_0 "asdf global terraform latest"
-exec_cmd_rc_0 "asdf list terraform"
-exec_cmd_rc_0 "sh -c 'terraform version'"
+if ! "$light"; then
+  echo "############### install terraform with asdf ###############"
+  if [ -z "$user" ]; then
+    exec_cmd_rc_0 "asdf plugin add terraform"
+    exec_cmd_rc_0 "asdf list all terraform"
+    exec_cmd_rc_0 "asdf install terraform latest"
+    exec_cmd_rc_0 "asdf list terraform"
+    exec_cmd_rc_0 "asdf global terraform latest"
+    exec_cmd_rc_0 "asdf list terraform"
+    exec_cmd_rc_0 "sh -c 'terraform version'"
+  else
+    exec_cmd_rc_0 "su - ${user} -c asdf plugin add terraform"
+    exec_cmd_rc_0 "su - ${user} -c asdf list all terraform"
+    exec_cmd_rc_0 "su - ${user} -c asdf install terraform latest"
+    exec_cmd_rc_0 "su - ${user} -c asdf list terraform"
+    exec_cmd_rc_0 "su - ${user} -c asdf global terraform latest"
+    exec_cmd_rc_0 "su - ${user} -c asdf list terraform"
+    exec_cmd_rc_0 "su - ${user} -c sh -c 'terraform version'"
+  fi
+fi
 
 # echo "############### install python3.12 ###############"
 # exec_cmd_rc_0 "sudo apt install -y python3.12 python3.12-venv"
@@ -207,18 +309,33 @@ exec_cmd_rc_0 "sh -c 'terraform version'"
 # exec_cmd_rc_0 "pyenv global 3.13"
 # exec_cmd_rc_0 "pyenv versions"
 
-echo "############### install python with asdf ###############"
-exec_cmd_rc_0 "asdf plugin add python"
-exec_cmd_rc_0 "asdf list all python"
-exec_cmd_rc_0 "sudo apt install -y build-essential libssl-dev zlib1g-dev libbz2-dev libreadline-dev libsqlite3-dev curl git libncursesw5-dev xz-utils tk-dev libxml2-dev libxmlsec1-dev libffi-dev liblzma-dev"
-exec_cmd_rc_0 "asdf install python latest"
-exec_cmd_rc_0 "asdf list python"
-exec_cmd_rc_0 "asdf global python latest"
-exec_cmd_rc_0 "asdf list python"
-exec_cmd_rc_0 "sh -c 'python --version'"
+if ! "$light"; then
+  echo "############### install python with asdf ###############"
+  if [ -z "$user" ]; then
+    exec_cmd_rc_0 "asdf plugin add python"
+    exec_cmd_rc_0 "asdf list all python"
+    exec_cmd_rc_0 "apt install -y build-essential libssl-dev zlib1g-dev libbz2-dev libreadline-dev libsqlite3-dev curl git libncursesw5-dev xz-utils tk-dev libxml2-dev libxmlsec1-dev libffi-dev liblzma-dev"
+    exec_cmd_rc_0 "asdf install python latest"
+    exec_cmd_rc_0 "asdf list python"
+    exec_cmd_rc_0 "asdf global python latest"
+    exec_cmd_rc_0 "asdf list python"
+    exec_cmd_rc_0 "sh -c 'python --version'"
+    # --- install git-remote-codecommit (for workshop-studio) ---
+    exec_cmd_rc_0 "pip install git-remote-codecommit"
+  else
+    exec_cmd_rc_0 "su - ${user} -c asdf plugin add python"
+    exec_cmd_rc_0 "su - ${user} -c asdf list all python"
+    exec_cmd_rc_0 "su - ${user} -c sudo apt install -y build-essential libssl-dev zlib1g-dev libbz2-dev libreadline-dev libsqlite3-dev curl git libncursesw5-dev xz-utils tk-dev libxml2-dev libxmlsec1-dev libffi-dev liblzma-dev"
+    exec_cmd_rc_0 "su - ${user} -c asdf install python latest"
+    exec_cmd_rc_0 "su - ${user} -c asdf list python"
+    exec_cmd_rc_0 "su - ${user} -c asdf global python latest"
+    exec_cmd_rc_0 "su - ${user} -c asdf list python"
+    # install git-remote-codecommit (for workshop-studio)
+    exec_cmd_rc_0 "su - ${user} -c pip install git-remote-codecommit"
+  fi
+fi
 
-echo "############### install git-remote-codecommit (for workshop-studio) ###############"
-exec_cmd_rc_0 "pip install git-remote-codecommit"
-
-sudo shutdown -r now
+if "$reboot"; then
+  sudo shutdown -r now
+fi
